added the quiz consition to allow the authorized users to create the quiz

+++ packages/nextjs/app/create-quiz/page.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import React, { ChangeEvent, FormEvent, useEffect, useState } from "react";
+import { useAccount } from "wagmi";
 // import { useRouter } from "next/navigation";
 import { withAuth } from "~~/components/withAuth";
 import { useScaffoldContractRead } from "~~/hooks/scaffold-eth";
@@ -23,6 +24,7 @@ const CreateQuizForm: React.FC = () => {
     answer: "",
     eventId: "",
   });
+  const { address: connectedAddress } = useAccount();
 
   const [editMode, setEditMode] = useState<string | null>(null);
   // const router = useRouter();
@@ -48,19 +50,22 @@ const CreateQuizForm: React.FC = () => {
 
   const getData = async () => {
     try {
-      const response = await fetch(`/api/quiz?id=${selectedEvent}`, {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-          "x-api-key": process.env.API_KEY || "",
-        },
-      });
+      const canAccess = checkQuizAccess();
+      if (canAccess) {
+        const response = await fetch(`/api/quiz?id=${selectedEvent}`, {
+          method: "GET",
+          headers: {
+            "Content-Type": "application/json",
+            "x-api-key": process.env.API_KEY || "",
+          },
+        });
 
-      if (response.ok) {
-        const data = await response.json();
-        setData(data.data);
-      } else {
-        console.error("Failed to fetch data");
+        if (response.ok) {
+          const data = await response.json();
+          setData(data.data);
+        } else {
+          console.error("Failed to fetch data");
+        }
       }
     } catch (e) {
       console.error(e);
@@ -72,64 +77,73 @@ const CreateQuizForm: React.FC = () => {
   }, [selectedEvent]);
 
   const handleSubmit = async (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (!formData.question || !formData.answer || formData?.options?.length < 3) {
-      return;
-    }
-    const correctAnswer = parseInt(formData.answer);
-
-    const newQuestion: Question = {
-      ...formData,
-      answer: formData.options[correctAnswer - 1],
-      eventId: selectedEvent,
-    };
-
-    if (editMode) {
-      // const updatedQuestions = questions.map(q => (q.eventId === editMode ? newQuestion : q));
-      // setQuestions(updatedQuestions);
-
-      const response = await fetch(`/api/quiz?id=${editMode}`, {
-        method: "PUT",
-        headers: {
-          "Content-Type": "application/json",
-          "x-api-key": process.env.API_KEY || "",
-        },
-        body: JSON.stringify(newQuestion),
-      });
-
-      if (response.ok) {
-        getData();
-        setFormData({
-          question: "",
-          options: ["", "", ""], // Reset options
-          answer: "",
-          eventId: selectedEvent,
-        });
-        setEditMode(null);
-      } else {
-        console.error("Failed to update question");
+    try {
+      e.preventDefault();
+      if (!formData.question || !formData.answer || formData?.options?.length < 3) {
+        return;
       }
-    } else {
-      const response = await fetch("/api/quiz", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          "x-api-key": process.env.API_KEY || "",
-        },
-        body: JSON.stringify(newQuestion),
-      });
+      const correctAnswer = parseInt(formData.answer);
 
-      if (response.ok) {
-        getData();
-        setFormData({
-          question: "",
-          options: ["", "", ""], // Reset options
-          answer: "",
-          eventId: selectedEvent,
-        });
+      const newQuestion: Question = {
+        ...formData,
+        answer: formData.options[correctAnswer - 1],
+        eventId: selectedEvent,
+      };
+
+      if (editMode) {
+        // const updatedQuestions = questions.map(q => (q.eventId === editMode ? newQuestion : q));
+        // setQuestions(updatedQuestions);
+        const canAccess = checkQuizAccess();
+        if (canAccess) {
+          const response = await fetch(`/api/quiz?id=${editMode}`, {
+            method: "PUT",
+            headers: {
+              "Content-Type": "application/json",
+              "x-api-key": process.env.API_KEY || "",
+            },
+            body: JSON.stringify(newQuestion),
+          });
+
+          if (response.ok) {
+            getData();
+            setFormData({
+              question: "",
+              options: ["", "", ""], // Reset options
+              answer: "",
+              eventId: selectedEvent,
+            });
+            setEditMode(null);
+          } else {
+            console.error("Failed to update question");
+          }
+        }
       } else {
-        console.error("Failed to create question");
+        const canAccess = checkQuizAccess();
+        if (canAccess) {
+          const response = await fetch("/api/quiz", {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+              "x-api-key": process.env.API_KEY || "",
+            },
+            body: JSON.stringify(newQuestion),
+          });
+
+          if (response.ok) {
+            getData();
+            setFormData({
+              question: "",
+              options: ["", "", ""], // Reset options
+              answer: "",
+              eventId: selectedEvent,
+            });
+          } else {
+            console.error("Failed to create question");
+          }
+        }
       }
+    } catch (error) {
+      console.error(error);
     }
   };
 
@@ -154,29 +168,40 @@ const CreateQuizForm: React.FC = () => {
 
   const handleDelete = async (id: string) => {
     try {
-      const response = await fetch(`/api/quiz?id=${id}`, {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-          "x-api-key": process.env.API_KEY || "",
-        },
-      });
-
-      if (response.ok) {
-        getData();
-        setEditMode(null);
-        setFormData({
-          question: "",
-          options: ["", "", ""],
-          answer: "",
-          eventId: "",
+      const canAccess = checkQuizAccess();
+      if (canAccess) {
+        const response = await fetch(`/api/quiz?id=${id}`, {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+            "x-api-key": process.env.API_KEY || "",
+          },
         });
-      } else {
-        console.error("Failed to delete question");
+
+        if (response.ok) {
+          getData();
+          setEditMode(null);
+          setFormData({
+            question: "",
+            options: ["", "", ""],
+            answer: "",
+            eventId: "",
+          });
+        } else {
+          console.error("Failed to delete question");
+        }
       }
     } catch (e) {
       console.error(e);
     }
+  };
+
+  const checkQuizAccess = () => {
+    const questionToEdit = eventData?.find((q: { mentorAddress: string }) => q.mentorAddress == connectedAddress);
+
+    if (!questionToEdit) {
+      throw new Error("You are not authorized to create quiz");
+    } else return true;
   };
 
   return (
@@ -264,7 +289,6 @@ const CreateQuizForm: React.FC = () => {
                 {data.map(
                   (
                     {
-                      eventId,
                       question,
                       options,
                       answer,
@@ -272,7 +296,7 @@ const CreateQuizForm: React.FC = () => {
                     }: { eventId: string; question: string; options: string[]; answer: string; _id: string },
                     index: number,
                   ) => (
-                    <li key={eventId} className="mb-4 p-2 border">
+                    <li key={index} className="mb-4 p-2 border">
                       <p>
                         <strong>Question {index + 1} :</strong> {question}
                       </p>
